refactor: Split page.tsx into seperate components
